backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.3
pydantic-settings>=2.5.0
feedparser>=6.0.11
python-dateutil>=2.9.0.post0
apscheduler>=3.10.4
google-generativeai>=0.8.3
python-dotenv>=1.0.1
